Implement the Annual calculator
At this stage, I'm not sure about the right abstraction, so I am doing a
bit of duplication while that comes out.

The simple and compounded daily and monthly interest formulas are
definitely different enough to make them their own thing, it makes the
calculators easy to test and to read, and it allows us to extend them
easily when we need to.

There's a high chance that I may not need to combine the calculator into
a single class. The only thing I foresee at the moment is potentially
having a class that selects which calculator to use, but I may end up
just using a small script to do that instead.