Use attribute readers instead of accessing directly instance variables
The benefit is quite small - if we determine that in the future we want
to have a reader do something other than just return the instance
variable, then there's very little refactoring to do.

This does not interfere with debuggers (as I originally observed) since
we're still setting environment variables and there's currently no
memoisation happening, which reduces visibility in debuggers.